+++ orderbook/src/api/create_order.rs
@@ -1,0 +1,118 @@
+use crate::storage::{AddOrderResult, Storage};
+use anyhow::Result;
+use model::order::OrderCreation;
+use std::{convert::Infallible, sync::Arc};
+use warp::{hyper::StatusCode, Filter, Rejection, Reply};
+
+const MAX_JSON_BODY_PAYLOAD: u64 = 1024 * 16;
+
+fn extract_user_order() -> impl Filter<Extract = (OrderCreation,), Error = Rejection> + Clone {
+    // (rejecting huge payloads)...
+    warp::body::content_length_limit(MAX_JSON_BODY_PAYLOAD).and(warp::body::json())
+}
+
+pub fn create_order_request() -> impl Filter<Extract = (OrderCreation,), Error = Rejection> + Clone
+{
+    warp::path!("orders")
+        .and(warp::post())
+        .and(extract_user_order())
+}
+
+pub fn create_order_response(result: Result<AddOrderResult>) -> impl Reply {
+    let (body, status_code) = match result {
+        Ok(AddOrderResult::Added(uid)) => (warp::reply::json(&uid), StatusCode::CREATED),
+        Ok(AddOrderResult::DuplicatedOrder) => (
+            super::error("DuplicatedOrder", "order already exists"),
+            StatusCode::BAD_REQUEST,
+        ),
+        Ok(AddOrderResult::InvalidSignature) => (
+            super::error("InvalidSignature", "invalid signature"),
+            StatusCode::BAD_REQUEST,
+        ),
+        Ok(AddOrderResult::Forbidden) => (
+            super::error("Forbidden", "Forbidden, your account is deny-listed"),
+            StatusCode::FORBIDDEN,
+        ),
+        Ok(AddOrderResult::PastValidTo) => (
+            super::error("PastValidTo", "validTo is in the past"),
+            StatusCode::BAD_REQUEST,
+        ),
+        Ok(AddOrderResult::MissingOrderData) => (
+            super::error(
+                "MissingOrderData",
+                "at least 1 field of orderCreation is missing, please check the field",
+            ),
+            StatusCode::BAD_REQUEST,
+        ),
+        Ok(AddOrderResult::InsufficientFunds) => (
+            super::error(
+                "InsufficientFunds",
+                "order owner must have funds worth at least x in his account",
+            ),
+            StatusCode::BAD_REQUEST,
+        ),
+        Err(_) => (super::internal_error(), StatusCode::INTERNAL_SERVER_ERROR),
+    };
+    warp::reply::with_status(body, status_code)
+}
+
+pub fn create_order(
+    orderbook: Arc<dyn Storage>,
+) -> impl Filter<Extract = (impl Reply,), Error = Rejection> + Clone {
+    create_order_request().and_then(move |order| {
+        let orderbook = orderbook.clone();
+        async move {
+            let result = orderbook.add_order(order).await;
+            if let Err(err) = &result {
+                tracing::error!(?err, ?order, "add_order error");
+            }
+            Result::<_, Infallible>::Ok(create_order_response(result))
+        }
+    })
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::api::response_body;
+    use model::order::OrderUid;
+    use serde_json::json;
+    use warp::test::request;
+
+    #[tokio::test]
+    async fn create_order_request_ok() {
+        let filter = create_order_request();
+        let order = OrderCreation::default();
+        let request = request()
+            .path("/orders")
+            .method("POST")
+            .header("content-type", "application/json")
+            .json(&order);
+        let result = request.filter(&filter).await.unwrap();
+        assert_eq!(result, order);
+    }
+
+    #[tokio::test]
+    async fn create_order_response_created() {
+        let uid = OrderUid([1u8; 56]);
+        let response = create_order_response(Ok(AddOrderResult::Added(uid))).into_response();
+        assert_eq!(response.status(), StatusCode::CREATED);
+        let body = response_body(response).await;
+        let body: serde_json::Value = serde_json::from_slice(body.as_slice()).unwrap();
+        let expected= json!(
+            "0x0101010101010101010101010101010101010101010101010101010101010101010101010101010101010101010101010101010101010101"
+        );
+        assert_eq!(body, expected);
+    }
+
+    #[tokio::test]
+    async fn create_order_response_duplicate() {
+        let response = create_order_response(Ok(AddOrderResult::DuplicatedOrder)).into_response();
+        assert_eq!(response.status(), StatusCode::BAD_REQUEST);
+        let body = response_body(response).await;
+        let body: serde_json::Value = serde_json::from_slice(body.as_slice()).unwrap();
+        let expected_error =
+            json!({"errorType": "DuplicatedOrder", "description": "order already exists"});
+        assert_eq!(body, expected_error);
+    }
+}
